Test nested data classes
